wallet: move EnsureKeyScope across all backends
